backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
psycopg[binary]==3.1.18
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.4
typing_extensions==4.15.0
